NXDRIVE-335: Use virtualenv for fonctional tests jobs
Several tree refactor has been made:
    - rename all requirements files to something unified (requirements-XXX.txt)
    - move tests folder from nuxeo-drive-client/nxdrive/tests to nuxeo-drive-client/tests

Bigs changes are:
    - we download and compile (if needed) our own Python, PyQt and dependencies versions
    - we do not use host softwares anymore (only basic ones like gcc or 7zip)

The process is the same on each platform, but we achieved it in different ways:
    - GNU/Linux and macOS: we use pyenv and compile what we need
    - Windows: we simply install Python version locally
The title is wrong, we do not use virtualenv as it is a mess on Windows and it happens to be a hell on other OS too.
Pyenv is the way to go and works like a charm, it is completely independant from the host.

For Jenkins jobs, we created 2 files in tools:
    - Jenkinsfile: to checkout, setup and launch the tests suite on all platforms
    - Jenkinsfile-release: to checkout and build releases on all platforms (GNU/Linux is not complete for now, but the process works)

Also:
    - Optimizations of imports
    - PEP8 changes here and there