NXDRIVE-849: Remove pairs without a parent
In case of a remote rename while syncing, it used to generate a lot of errors
because of an inexisting parent folder. The proposed solution remove the doc_pair
from the DTB and, if needed, it will be redetected by the next local scan with
up-to-date informations. If not, either the document is now up-to-date
or it will be processed the next, ... , next local scan.

Also:
    - Filter out *.nxpart from processors
    - Rework how reasons are determined in JS
    - Convert Workers to more generic ones using **kwargs
    - Remove Wachdog sentinel
    - Set file logging level to DEBUG by default
    - Locales are sorted in GUI

NXDRIVE-626: Review translator.py, its tests and sort i18n.js to facilitate translations
NXDRIVE-863: Enhance analytics with the locale, FS encoding and a better OS detection