NXDRIVE-873: Fix report generation with a standard (non administrator) user on Windows
The error was due to the write of debug.log in a directory that the
current user should not have write rights.  I reviewed the entire
class with better exceptions and several sentinels to prevent the
bug report to crash.  If one file fails, then just print the error
and check the next one.