NXDRIVE-897: Handle remotes ignored files
If one document is created server side with an ignored extension,
it will be ignored client side with an informative message.

On the other direction, if a document is created with an ignored
extension, it will not be uploaded server side. An informative
message may or may not be displayed depending which method handle
it.