NXDRIVE-897: Right management of AutoCAD behaviors
Handle temporary files
----------------------

By default, ignore .dwl and .dwl2 extensions (used by AutoCAD), also
.nxpart (Drive download suffixe).  Another temporary files created by
AutoCAD are "atmpXXX" (where XXXX are numbers) and we have no other
solution, for now, but to blacklist that file name pattern.

Handle remote ignored files
---------------------------

If one document is created server side with an ignored extension,
it will be ignored client side with an informative message.

On the other direction, if a document is created with an ignored
extension, it will not be uploaded server side. An informative
message may or may not be displayed depending the called method.