NXDRIVE-898: Add a system tray context menu (#267)
A lot of GUI improvements here.  We now have a systray menu that is
full working as expected on all OSes.  I added a context menu on
GNU/Linux and Windows, and the midde click opens settings directly.
On macOS, the behavior does not change at all.

There are several fixes for the WIP NXDRIVE-870: the internal API
changed in many ways, refer to docs/technical_changes.md for more
informations.


NXDRIVE-711: System tray menu acts weird

On GNU/Linux and macOS, when you clicked on the systray icon, it opened
the menu.  Then, you clicked outside the menu and it changed nothing.
You had to second click elsewhere to close it.


NXDRIVE-694: Systray menu: needs double click to activate

On GNU/Linux, to have access to anything inside the menu, you had to
click twice.