NXDRIVE-404: Handle custom SSL certificates
A message box will display the custom certificate
and ask the user what to do next. If it is decided
to bypass SSL verifications, then it will apply
everytime for the impacted account.

Note that we do not check for bad certificates
*after* the account creation as it is unlikely
to change over the time. It is more important
to test at the beginning. If such scenario
happens, the user will have to tweak SSL
behavior with ca-bundle or ssl-no-verify options.