NXDRIVE-1497: Fix alert when a server is not responding
This commit unifies the way server availability and compatibility is checked while adding an account or when polling for updates.
It adds a Login class that extends the Flag enum class. This way, the Login flow available (OLD, NEW or UNKNOWN) can easily be tracked, combined when using multiple accounts, and follow-up decisions are clearer.