NXDRIVE-1535: Handle Unauthorized errors from Processor
Instead of stopping the synchronization and putting
the Engine in offline mode, as this error can
happen on different level in the tree, we will
put those documents in error.

And when the user will have enough rights to sync
them, simple click on the retry button for those
errors will automaticallysync them.

If the user lost its token for real, the
RemoteWatcher will set invalid credentials when
looking for remotes changes.