NXDRIVE-1436: Split test files
This will speed-up the overall testing time.

First move to a complete uniquetestenv and
possible parallelization.

Created several "tests" subfolders:

* functional: new format, pytest only, of tests
    requiring a Nuxeo instance
* old_functional: tests not yet converted to the
    new framework
* unit: simple code tests, no Nuxeo instance required

Tests that are in "functional" or "unit" folders
will be ran in parallel using pytest-xdist.

The next step is to think how to generate unique
"remote" and "local" clients (with specific workspaces)
to be able to migrate old functional.

The parallelization is possible because now we
can instanciate more that one Manager. I removed
the singleton restriction.
We can also use the Manager as a context manager
to automatically free resources at exit.

---

This PR also fixes several issues found while
converting tests:

* FolderAlreadyUse exception, when binding an
    account, was not checking correctly the path equality
* Options.mock() was raising a TypeError when the tested
    function had arguments
* Reworked Engine._add_top_level_state() with better
    resource handling