NXDRIVE-1493: Add DirectEdit support for the Adobe Creative Suite
* Analyse Photoshop behavior (same for Illustrator)

* Move `Item` and `Items` to `objects.py` for further needs

* Refactor `DirectEdit.handle_watchdog_event()`

Simply move out the try...except block outside
the method to have less indentation.

* Introduce `osi/windows/files.py`

It is used to add application specific checks on
opened files. It currently handles Photoshop and
Illustrator.

* Add comments in `autolocker.py`

* Refactor `Manager.open_local_file()`

Created OS specific `open_local_file()` methods instead of a big one.
It is now easier to follow and easier handle errors specific to the OS.

Introduce `osi/linux/linux.py` for the occasion.

* Introduce `osi/linux/files.py`

* Use the new `get_other_opened_files()` to catch Adobe CC behaviors

For now, it is referencing `osi/darwin/files.py` which is not yet finished.

* Lock the document at opening

* Add `osi/darwin/files.py`

We can use AppleScript to retrieve opened document
in Photoshop and illustrator. Pretty cool!

* Refactor `osi/*/files.py`

Added utils.py::`compute_fake_pid_from_path` to
compute the CRC32 of a given path. This is not
the most secure but we will unlikely to have
collisions as it will be in use for at most
3 files at the same time (seeing big).

* Update the investigation document to include macOS snippet