NXDRIVE-1467: Implement FS decorator on Windows
This includes having a TCP socket server in Drive to listen and answer with a status to DLLs. These DLLs are based on the Liferay Nativity (https://github.com/liferay/liferay-nativity) project.

* NXDRIVE-1485: Properly display and refresh icons with FS decorator on Windows

The method used by Drive to read the socket was triggering a close of the connection.

* NXDRIVE-1538: Package and test FS decorators on Windows 7 and 10

Build the DLLs on Windows 10, bundle them in an addon installer as well as in the system-wide installer.

* NXDRIVE-1544: Automate Windows FS decorator DLLs build

Be able to build a new set of DLLs from the usual deployment script.