NXDRIVE-1501: Prompt for confirmation when a user deletes a document locally
The default behavior when detecting a file deletion is now to unsynchronize the document. It is filtered and can be synchronized again by checking them in the Filters folder.
It can be changed in the settings window to the previous behavior that trashes the file on the server.
Furthermore, the user will be prompted to validate or rollback a deletion every time unless he checks a box to not be prompted again.