NXDRIVE-986: Creating a new folder then renaming generates meaningless URLs
* Introduced Remove.move2() to call Document.Move.

The trick here is to call move2() inside the same parent
but with a new name. It will update the name and so the URL.

WARNING: this is a big change as it will invalidate any
shared URL containing the document's name. In other words,
if one shared a simple URL, and if the document is renamed
by anyone, the URL will end on a 404 error (not found).
This is not the case if one used the official mecanism and
shared the document using the "Share link" that is using
the document's UID.