NXDRIVE-1578: DirectEdit'ing a doc and saving multiple times by second fails
The fix is to skip the comparison with the blob on the server
when the document is being DirectEdit'ed: as it should be locked,
the document cannot be modifed by anyone else.
This check was creating several conflicts (that were not displayed
because of another bug) and finally the original version was restored
when the DirectEdit'ion was over.

As a bonus, we save 2 HTTP calls: now only 1 call for each
and every modification!

* Introduce DirectEditDetails to simplify DirectEdit._extrac_edit_infos() usage.
* Fix message box not displayed in Application._direct_edit_conflict().
* Move out an if condition outside try...except in RemoteWatcher._update_remote_states().