NXDRIVE-695: Update all DB local paths on FS move event
This fixes the issue on macOS and Windows.

On GNU/Linux, the deleted event is against the old
"/folder/subfolder/file.txt" instead of
"/folder-renamed/subfolder-renamed/file.txt".

This is a corner case and I have no idea how to fix it.
So the fix is postponed on GNU/Linux.
To follow with NXDRIVE-1690.