NXDRIVE-1219: Use a dedicated temporary folder for downloads
We were downloading files into temporary hidden files.
It was causing a lot of false FS events, generating
significative noise, thus complexifying the synchronization.

We are now using a dedicated temporary folder. It must be
on the same partition of the local sync folder and outside
the local sync folder. The first condition is not yet checked
and will be tackled with NXDRIVE-1779.

To generate less noise while downloading a file, I introduced
a new integer option, tmp_file_limit (defaults to 10, this is
the file size in MiB). Files smaller than this limit will be
written at once to the temporary file rather than chunk by chunk.

Benefits:
 - no more false FS events;
 - no more temporary .nxpart files;
 - files smaller than tmp_file_limit will be
   downloaded in one go in-memory, reducing
   disk I/O.

Based on the original patch from Léa Klein (@KLea).