NXDRIVE-1734: Ignore decoding issues when retrieving the path stored in xattrs
Introduced the LocalClient.set_path_remote_id()
static method to be able to test the feature
without needing an account.

While testing, I saw that on Windows, when a
file is not found and we want to set the remote
ID, an exception would be rasied. But not on
other OSes. So I aligned the behavior.