NXDRIVE-1775: Bypass PermissionError raised from Path.resolve()
Pathlib.Path.resolve() will raise a PermissionError
when the file is opened elsewhere. This is not
convenient as we only want to get the full path
(without any action).

We could fallback on using pathlib.Path.absolute()
is that case. This may be problematic when using
symlinks, but as of now there is no such files in Drive.

Patch applied to LocalClient.get_path() and normalized_path().