NXDRIVE-1732: Better handling of the Direct Edit protocol
The regexp to parse the DirectEdit URL is now stricter:
the download URL must be part of it. It is a requirement
since a while, but the regexp was not updated.

Also, the same regexp has been improved to skip an anonymous
group around the document's UID.

Tests have been added for parse_protocol_url().