NXDRIVE-1783: Handle account addition with already used local folder
This happens in that scenario:
 - Add a new account using the local folder "/home/USER/drive".
 - Delete the Engine database.
 - Add a new account using the local folder "/home/USER/drive".

FolderAlreadyUsed is raised instead of deleting the old database entry because
it is convenient to be able to restore the DB file later for whatever reason.
And there is no popup to ask the user for its deletion for the same purpose.
Note that the use case is rare enough to be handled that way. In fact, in years,
that happened only while testing the application a lot.