NXDRIVE-1803: Prevent unexpected downgrades
When using the centralized channel and the desired
version is not set and the current version is
from another channel than the release one, a
downgrade is done. This should rather display
a popup asking the user to either downgrade
on the release channel or continue using the
channel from the current version.