NXDRIVE-1839: The GUI should be responsive when syncing a lot of files
- Added timers to control when the files list in
the systray can be updated. Same for syncing
docs and synced docs counts.

Some refactoring in view.py on how rows insertion
and deletion are done. The new code is cleaner
and less resources greedy.

There is a small downside though: the remaining
docs to sync (the last iteration) is not updated
until the sync is actually over. But this is
another issue not related to the subject and
should be tackled another time.

- Unchunked transfers are now lighter in their
traitment: downloads have no more DownloadAction
nor VerificationAction.