NXDRIVE-1856: Prevent duplicate creation via Direct Transfer
Before uploading anything, a check is to be done to prevent duplicate creations.

By default, if the document does not exist: create a new document.

But if a document with the same name already exists on the server:

Case 1: The document has a blob stored into the file:content xpath, asks the user for:
  1) upload cancellation;
  2) blob replacement.

Case 2: The document has no attached blob (in case of a previous failure at the upload step):
  - upload and attach the blob.

---

Note about possible duplicate creation via a race condition client <-> server.
Given the local *file* with the path "$HOME/some-folder/subfolder/file.odt",
the file name is "file.odt".

Scenario:
    - Step 1: local, check for a doc with the path name "file.odt" => nothing returned, continuing;
    - Step 2: server, a document with a path name set to "file.odt" is created;
    - Step 3: local, create the document with the path name "file.odt".

Even if the elapsed time between steps 1 and 3 is really short, it may happen.

What can be done to prevent such scenario is not on the Nuxeo Drive side but on the server one.
For now, 2 options are possible but not qualified to be done in a near future:
    - https://jira.nuxeo.com/browse/NXP-22959;
    - create a new operation `Document.GetOrCreate` that ensures atomicity.