NXDRIVE-1857: Handle multi-account for Direct Transfers
A popup with accounts is shown first when there
is more than 1 account registered.

If the user hit ESC or click on Cancel, the upload
is aborted.

When hitting ENTER or clicking on OK, the selected
account is used for the Direct Transfer. There is
no sorting on accounts yet, the first printed will
be used by default.