NXDRIVE-1579: Disable synchronization upon server configuration
Introduced a new option: synchronization_enabled (defaults to True).

When that option is set to False, Local and Remote whatchers will be
started but nothing will be synced at all; Direct Edit and Direct Transfer
features will continue to work as expected. I tried to be the less
intrusive as possible.

I duplicated tests for those 2 features: one class in "normal"
mode, when sync is effective; and one class with sync disabled.