NXDRIVE-1939: Use a temp dir located on the same drive as the local folder in LocalClient.rename()
Using the download folder from the engine will
ensure both files are on the same drive.

[CI] Our Windows agents were updated to create
a second partition available on "Q:\", it is
customizable with the `SECOND_PARTITION` envar.
See NXBT-3144 for details.

I also renforced `.is_case_sensitive()` for when
`%TEMP%` is restricted on Windows. And removed its
call from `__init__()` for a lazy check. Before
those changes, Drive was unusable; now it will start.