NXDRIVE-1951: Do not try to update contents of a folderish document
Folderish documents were not properly checked on remote modifications.
So the scenario of trying to update folderish contents was happening.

The fix is to filter out those documents.

Another smaller fix was applied to not get the size of a folder.
Even if it is a valid operation, it is useless and not required.