NXDRIVE-1981: Better improvement patch for safe_filename()
The benchmark used for the original patch was not realistic and
results were wrong. I added a micro-benchmark suite to ensure
the current implementation is the best one as of now.

Introducing unicode characters in filenames make results even worse.
But `str.replace()` still has the best score. The code is cleaner
and easier to read.

Also removed `safe_os_filename()`.

Filename sanitization is now done equally on all OSes.