NXDRIVE-1976: [macOS] Do not fail the auto-update on unmountable volume
The last step of the auto-update is failing on some Catalina machines
due to the new security framework. It is about unmounting the DMG
volume. From the user POV, it always worked, the new version was effective
at the next start.

We can "safely" ignore that error as the update actually did work.
Having smooth updates is the priority.