NXDRIVE-1918: Fix batch retrieval with the S3 upload provider
Setting the `file_idx` kwarg when retrieving a batch is
required for the default upload provider only.
For S3, there is no such file at any index until the S3
upload itself is completed and the call to
`batch.complete()` is done.

The commit contains various small adjustments to finish
the S3 integration.