NXDRIVE-2010: Restart the download on HTTP 416 error (range not satisfiable)
When a download is paused, it may happen that the file size on the server
is modified. In this particular case, an HTTP 416 error will be raised.
We want the local temporary file to be deleted and the download to be
restarted.
The processor.py file have been updated to catch and process the error.
A new test have been added in the test_synchronization.py file.
Also the changelog have been updated.