NXDRIVE-2079: Special case handling when testing the auto-updater
Our packaging job tests the auto-updater before validating
an alpha release. For that, it does not use an account and bypass
that check. But with d8bf965 the auto-updater is started only
after the server's config is fetched, which cannot happen without
account.

So we add a special case check to make the alpha release job happy again.