NXDRIVE-2113: [Direct Edit] Add a warning on upload when server-side lock is disabled
Uploading a new blob of an ongoing Direct Edit'ed document should be made only when the
user had already locked it.
The auto-locking of documents can be disabled on Drive throught an option, exposing
the user to possibles concurrent updates of the document.

A warning log has been added when a Direct Edit is done while this option is activated.