NXDRIVE-2132: Do not allow Direct Edit on proxies
Direct Edit on proxies should be forbidden as it causes
many errors that fills the logs.

The document download is now aborted and a graphical error
is displayed when the user try to use Direct Edit on a proxy.
The translation file has been updated with the new error message.
A functionnal test has been added.
Also changelog has been updated.