NXDRIVE-2131: [Direct Edit] Handle HTTP 413 error: Request Entity Too Large
When trying to upload changes in Direct Edit, and if the target is a proxy, then
the `HTTPError` 413 will be raised, filling server logs, client logs and
Sentry events. The patch is a completion of NXDRIVE-2132.

The error is now catched and the upload skipped when this error is raised.