NXDRIVE-2129: [Direct Edit] Orphaned documents should first be remotely unlocked before being cleaned up
When the application is shut down during a direct edit, the edited file
stay locked on the remote server and the temporary file is not deleted.
When starting, the expected behaviour is to unlock the document on
the remote then remove the local orphan file.

This commit fix the current behaviour in which Drive fails to retrieve
correctly the informations about the locked files from the database.
This prevent Drive to recognize orphan local files as locked.
This is the second step of a three steps fix of the bug.
Also changes has been updated.