NXDRIVE-2129: Orphaned documents should first be remotely unlocked before being cleaned up
When the application is shut down during a direct edit, the edited file
stay locked on the remote server and the temporary file is not deleted.
When starting, the expected behaviour is to unlock the document on
the remote then remove the local orphan file.

This commit add a functionnal test on the direct edit system for
unlocking orphan files.
This is the last step of a three steps fix of the bug.