NXDRIVE-2193: Stop trying to guess the server URL
When adding an account the app was trying to understand
the given URL by deconstructing and reconstructing it using
different patterns. This behaviour leaded to several issues
and was a poor end user feature in the end.
From now, a simple syntax check will be done on the URL and if
it is wrong then the user will have to fix it.

Changelog has been updated.