NXDRIVE-2219: Display a remote folder icon for each transfer
This fixes UI performance issues because there is no need
to do QML > C > Python > C > QML to fetch the remote
folder path and UID (it was displayed in the header
of the Direct transfer window).

Another performance issue was tackled: the use of xattrs to store
remote paths and UIDs. Sending 50,000 files was freezing the UI;
not it is still frozen but quite less. I may need additionnal work
to completely tackle that issue.

Some interesting fixes/improvements:

- A nice folder icon is shown on the left of each progress bar
  to access the remote destination folder.
- Clean-up Direct transfer tests.
- Notifications are shown when the uploaded file is >= 25 MiB.
- The buttons available when cancelling a transfer are now Yes/No.
- The window is kept showed when there are transfers to handle but
  not directly available at the time (because of I/O access maybe).