NXDRIVE-2231: Partly revert the patch
We finally can't rely on ES for that task. ES is intrinsically async
and the query does not make sens anymore then. Restored the old behavior.

Also fixed the duplicate creation (the `overwite` parameter).

Another small improvement was done: if `replace_blob` is True, then no
need to check for existent documents on the server.