NXDRIVE-2309: [Direct Transfer] Implement Active Transfer sessions tab
We want to be able to interact with the Direct Transfers sessions
throught the GUI. Two new tabs have been added to the Direct
Transfer Window. These tabs list the ongoing and completed sessions
with some usefuls informations about them.
It is now possible to pause / resume / cancel a Direct Transfer session.