NXDRIVE-2349: Fix a QML margin in the account addition
Since the upgrade to PyQt 5.15.1, when adding an account there is
a missing margin above the URL field. But as soon as a character
is entered into the field, the margin comes again and all is fine.

So let's increase the margin and see later when the Qt/PyQt issue
will be addressed to remove it.