NXDRIVE-2368: Bump the chunk_limit upper limit from 20 MiB to 5120 MiB (5 GiB)
The original limit (20 MiB) was quite a random value. This may be inefficient
to use such low value when one can upload a part size of 5GiB to S3 and test
a local network with terabyte links.

So I bumped the upper limit to 5120 MiB to allow using S3 at its full power.