NXDRIVE-2309: Fix a SQL syntax error
For one reason or another, it happens that the `completed_on` field
of a session is set while at least one transfer is not finished.
This seems something related to concurrency at hight speed, so why
we didn't see it before.

    Retry locked database #2, args=('UPDATE Sessions SET uploaded = ?, status = ? , completed_on = 2020-11-06 13:58:25 WHERE uid = ?', (13, 4, 21)), kwargs={}

    Traceback (most recent call last):
      File "nxdrive/engine/dao/sqlite.py", line 147, in execute
    sqlite3.OperationalError: near "13": syntax error

Anyway, I simplified the logic to not rely on f-strings to prevent the SQL error
and to update the `completed_on` field only when revelant, i.e.: when the session is
actually completed.