NXDRIVE-2379: Review how the GUI is repainted to fix a crash
Previously, items were added and removed from the ListView
when a change in sessions or dt_items was detected.
This behaviour caused the app to crash if an item
from the list was removed while in use (cancel window).

Now, a fixed list of items is used an just edited when
changes are detected, preventing the interface to
remove used items.