NXDRIVE-2270: [macOS] Big Sur support
This is a commit ahead of the time: it will make sens only when
we will be using Python 3.8.6+ and PyInstaller 4.1+.

But it has been verified that using those versions + this patch
make the app working on macOS Big Sur.

Note: we still can't use Big Sur as a development machine due to
issues on other modules (`packaging`), Python and Qt.