NXDRIVE-2309: [Direct Transfer] Improve the session pause behavior
Improvements following 3baefb2fe806ced85226976c715b1fbe4bf4ad7c.

Before the patch, all items to be transferred were started and then
paused when their session was paused. This was quite bad because as
a transfer was initiated, a call to the server was made for at least
retrieve a `batchId`. Even worse, all remaining transfers not yet handled
were started _at once_; resulting in N parallel calls to the server and
entries in the `Uploads` database table.

Now, when pausing a session, ongoing transfers will be finished for non-chunked
ones, the others will be paused ASAP.

Another issue was fixed when starting the app and a session is paused: all its
transfers were added to the queue and then filtered out by the `Processor`.
The SQL query was improved to prevent pushing such pairs into the queue.

As always: less is better, let's be lazy :)