NXDRIVE-2329: [Direct Transfer] Add a button to create a folder within the upload flow
A new button has been added in the Direct Transfer options. This
button allow the user to create a new folder on the selected
remote path. Selected files and folder will be uploaded inside
the new remote folder it this one is defined in the options.

Also changelog has been updated.