NXDRIVE-2472: Prefer to not setting the protocol (#2368)
Python will automatically use the most secure protocol by default.

The security issue is no more valid because since Python 3.6, the best protocol
is used based on the OpenSSL version. And we are shipping a desktop app with
one given version of OpenSSL which is always the latest shipped with the Python
version in used. It is then secure.