NXDRIVE-2140: Handle documents with non-standard or empty digest
From now, all documents will be kept in the database.
Documents with non-standard digest will be marked as
ignored with the appropriate description to help users
understanding why.

For some documents, their digest is or will be computed asynchronously
on the server. When the digest will be available, an Audit event will
alert us and then:

- The digest will be updated in the database.
- The document will be unmarked as ignored.
- The process will resume and the document will be synchronized.

For other documents where the digest is problematic, like Live Connect
documents, they will be kept around "forever". It will allow us to potentially
change the way we want to handle them in a future release.