NXDRIVE-2526: Light icon option is not saved anymore (#2478)
With NXDRIVE-2513, a check in Manager.set_config() was added to prevent
updating the database when values are the same. But it does not work when
the option is not registered.

We simply registedred a new option, and it will be convenient as now the option
can be set from the local or remote configuration.