NXDRIVE-2523: Prevent crashes because of too many open file descriptors
The issue was not visible quickly but when Drive was running a long
time with a lot of synchronization actions. Each thread was opening
a connection to the SQLite database file, but it was never released.

It was kind of weird to manager connections like that. Now, read/writes
cursors are well handled. It results in a slower synchronization when
it is a matter of thousands small files.
We decided to go with the stability over the velocity.