NXDRIVE-2321: Fix file name escaping when checking for document existence on the server
The patch prevents NXQL errors to happen again when a filename contains a carriage return character.

But it will not really prevent creating duplicates because the check will be done on the local filename
and a similar document would unlikely have a carriage return in its own `dc:title`. So the request is valid
but the check is useless for such names.
For instance, at the time writing those lines, a document named `My \r file.patch` will be save
as `r file.patch` when using the `FileManager.Import` operation.

I am not sure how to handle those documents, and if we want to add complexity for such a corner-case.