NXDRIVE-2694: Review handling of sync roots (#2920)
NXDRIVE-2694: Review handling of sync roots

Sync roots will be renamed according to their `N` parents' names.
`N` defaults to 2 and can be controled through the `sync_root_max_level`
option.
The option can get up to 4 levels or hierarchy. Set it to 0 to disable
the new behavior.

Sync root names are also always shortened to 47 chars (ellipsis will be
included in the middle when it was altered).